+++ src/Frontend/AppAppoef/app/src/main/AndroidManifest.xml
@@ -15,23 +15,23 @@
         android:theme="@style/Theme.AppAppoef"
         tools:targetApi="31">
         <activity
-            android:name=".OrientacaoProvaActivity"
+            android:name="org.appoef.appappoef.OrientacaoProvaActivity"
             android:exported="false" />
         <activity
-            android:name=".InformacoesAssociacaoActivity"
+            android:name="org.appoef.appappoef.InformacoesAssociacaoActivity"
             android:exported="false" />
         <activity
-            android:name=".ProvaActivity"
+            android:name="org.appoef.appappoef.ProvaActivity"
             android:exported="false" />
         <activity
-            android:name=".PrincipalActivity"
+            android:name="org.appoef.appappoef.PrincipalActivity"
             android:exported="false"
             android:label="@string/title_activity_principal" />
         <activity
-            android:name=".CadastroActivity"
+            android:name="org.appoef.appappoef.CadastroActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name="org.appoef.appappoef.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
