+++ src/Frontend/AppAppoef/app/src/main/java/org/appoef/appappoef/CadastroActivity.java
@@ -1,4 +1,4 @@
-package com.example.appappoef;
+package org.appoef.appappoef;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -19,6 +19,7 @@ import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.Volley;
+import org.appoef.appappoef.R;
 
 import org.json.JSONException;
 import org.json.JSONObject;
