+++ src/Frontend/AppAppoef/app/src/main/res/layout/activity_cadastro.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CadastroActivity"
+    tools:context="org.appoef.appappoef.CadastroActivity"
     android:background="@drawable/background_cadastro">
