+++ src/Frontend/AppAppoef/app/src/main/res/layout/activity_orientacao_prova.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OrientacaoProvaActivity">
+    tools:context="org.appoef.appappoef.OrientacaoProvaActivity">
 
     <ScrollView
         android:layout_width="0dp"
