+++ src/Frontend/AppAppoef/app/src/main/res/layout/activity_principal.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PrincipalActivity"
+    tools:context="org.appoef.appappoef.PrincipalActivity"
     android:background="@drawable/background_tela_principal">
